角色更新功能

+++ testdb/dao/permissionDao.go
@@ -378,7 +378,7 @@ func RetrieveRoleData(fetchDataBody *models.Role) (dataResBody []models.Role, to
 	return results, totalCount, err
 }
 
-func DeleteRole(roleId int, tx *sql.Tx) (err error) {
+func DeleteRole(roleId int64, tx *sql.Tx) (err error) {
 	_, err = tx.Exec("DELETE FROM `tb_roles` WHERE role_id = ? ",
 		roleId)
 	if err != nil {
@@ -394,7 +394,7 @@ func DeleteRole(roleId int, tx *sql.Tx) (err error) {
 	return
 }
 
-func DeleteRolesAndFunctionsByRoleId(roleId int, tx *sql.Tx) (err error) {
+func DeleteRolesAndFunctionsByRoleId(roleId int64, tx *sql.Tx) (err error) {
 	_, err = tx.Exec("DELETE FROM `tb_roles_functions` WHERE role_id = ? ",
 		roleId)
 	if err != nil {
